More gles fixes

+++ base/shaders/mesh_posnortex.frag.glsl
@@ -16,7 +16,7 @@ void main() {
 	float occlusion = 1.0;
 	n /= (abs(n.x) + abs(n.y) + abs(n.z));
 	n.xy = n.z >= 0.0 ? n.xy : octahedron_wrap(n.xy);
-	uint matid = 0;
+	uint matid = uint(0);
 	frag_color[0] = vec4(n.xy, roughness, pack_f32_i16(metallic, matid));
 	frag_color[1] = vec4(basecol, occlusion);
 	vec2 posa = (wvpposition.xy / wvpposition.w) * 0.5 + 0.5;
